requirements.txt
torch>=2.0.0
torchvision>=0.15.0
Pillow>=10.0.0
albumentations>=1.3.0
onnxruntime>=1.15.0
onnxruntime-gpu>=1.15.0
pyyaml>=6.0
tqdm>=4.65.0
tensorboard>=2.13.0
editdistance>=0.6.0
numpy>=1.24.0
